requirements.txt
# ─────────────────────────────
# Core Web Framework
# ─────────────────────────────
fastapi==0.115.0
uvicorn==0.30.1
starlette>=0.37.2,<0.39.0
aiofiles==23.2.1
jinja2==3.1.4

# ─────────────────────────────
# NLP & Text Processing
# ─────────────────────────────
spacy==3.7.4
nltk==3.9.1
PyPDF2==3.0.1
python-docx==1.1.2
docx2txt==0.8
regex==2023.12.25

# ─────────────────────────────
# Machine Learning / AI
# ─────────────────────────────
numpy==1.26.4
scikit-learn==1.5.1
transformers>=4.41.0,<5.0
sentence-transformers==5.1.0
torch>=2.3.0

# ─────────────────────────────
# Azure & OpenAI API
# ─────────────────────────────
openai==1.47.0
azure-identity==1.17.1
azure-storage-blob==12.22.0
python-dotenv==1.0.1

# ─────────────────────────────
# Reporting / File Export
# ─────────────────────────────
openpyxl==3.1.5
reportlab==4.2.2
pandas==2.2.3

# ─────────────────────────────
# Testing / Utilities
# ─────────────────────────────
pytest==8.3.3
httpx==0.27.2

# ─────────────────────────────
# Optional tools for DevOps / CI
# ─────────────────────────────
black==24.8.0
isort==5.13.2
flake8==7.1.1


# ─────────────────────────────
# Authentication
# ─────────────────────────────
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
requests==2.31.0
cachetools==5.4.0
